pyproject.toml
[project]
name = "bcog-final-project-dino-game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pygame-ce>=2.5.7",
]

[tool.uv.workspace]
members = [
    "src",
]
